Fix Build
I ran `npm install` with the following versions and received an error.

  npm: 9.6.7
  node: v20.3.1

The error is explained in
[this](https://stackoverflow.com/questions/69692842/error-message-error0308010cdigital-envelope-routinesunsupported)
stack overflow post and is related to upgrading past node v16. I ran
`npm audit fix --force` which didn't actually fix the issue, but then I
upgrade the version of `@vercel/ncc` and the build worked after that.

I decided to upgrade all the dependencies as well since I updated one.