Test that the working-directory option is working
This adds a new top-level directory: `test-code-2/`.

This directory contains a `Main.hs` Haskell file with poor formatting.
If we were to naively run `run-fourmolu` on this Haskell file, then the
Action would detect that some changes need to be made to the file, and
fail.

However, this commit also adds the `working-directory` option with a
value of `test-code/`, so the `run-fourmolu` action won't see the
`test-code-2/Main.hs` file.

We expect that the `run-fourmolu` action succeeds in this case.